The passenger home is ready
