+++ HSL.V2/Views/SearchPassengerView.swift
@@ -9,96 +9,112 @@ import SwiftUI
 
 struct SearchPassengerView: View {
     // The `showSearchScreen` property is a binding that indicates whether the search screen should be shown or hidden.
-    @Binding var showSearchScreen: Bool
+    @Binding var showSearchView: Bool
     
     @StateObject private var viewModel = PassengerViewModel()
     
-    
-    
-    @State private var isDragging = false
-    
-    @State var busNum: String = ""
-    
-    @State private var currentHeight: CGFloat = 200
-    
-    let minHeight: CGFloat = 100
-    let maxHeight: CGFloat = 700
-    
     var body: some View {
         ZStack(alignment: .bottom){
-            if showSearchScreen {
+            if showSearchView {
                 // Show the search screen if showSearchScreen is true
-                mainView
-                    .transition(.move(edge: .bottom))
-            }
-        }
-        .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .bottom)
-        .ignoresSafeArea()
-        .animation(.easeInOut)
-        
-    }
-    
-    var mainView: some View {
-        VStack {
-            ZStack{
-                Capsule()
-                    .frame(width: 40, height: 6)
-            }
-            .frame(height: 40)
-            .frame(maxWidth: .infinity)
-            .background(Color.white.opacity(0.00001))
-            .gesture(dragGesture) // apply the gesture to the main view
-            
-            ZStack {
                 VStack {
                     searchBus(searchText: $viewModel.searchText)
-                        .padding(.bottom, 50)
-//                    Form {
-//                        List {
-//                            Text("To A")
-//                            Text("To B")
-//                        }
-//                    }
+                        .transition(.move(edge: .bottom))
                 }
-                .frame(height: currentHeight)
-                .frame(maxWidth: .infinity)
             }
-            .animation(isDragging ? nil : .easeInOut(duration: 0.45))
         }
     }
-    // The `prevDragTranslation` property is used to keep track of the previous drag translation.
-    @State private var prevDragTranslation = CGSize.zero
     
-    var dragGesture: some Gesture{
-        DragGesture(minimumDistance: 0, coordinateSpace: .global)
-            .onChanged { val in
-                if !isDragging {
-                    isDragging = true
-                }
-                let dragAmount = val.translation.height - prevDragTranslation.height
-                if currentHeight > maxHeight || currentHeight < minHeight {
-                    currentHeight -= dragAmount / 6
-                } else {
-                    currentHeight -= dragAmount
-                }
-                prevDragTranslation = val.translation
+}
+struct searchBus: View {
+    @Binding var searchText: String
+    @ObservedObject private var viewModel = PassengerViewModel()
+    
+    var body: some View {
+        VStack {
+            HStack {
+                Image(systemName: "magnifyingglass")
+                    .opacity(searchText.isEmpty ? 1.0 : 0.0)
+                    .foregroundColor(Color.theme.blue)
+                TextField("Search by your bus number...", text: $searchText)
+                    .disableAutocorrection(true)
+                    .frame(height: 45)
+                    .background(Color(.systemGroupedBackground))
+                    .onChange(of: searchText) { searchText in
+                        viewModel.searchText = searchText
+                    }
+                    .overlay(
+                        HStack {
+                            Image(systemName: "magnifyingglass")
+                                .offset(x: 10)
+                                .opacity(searchText.isEmpty ? 0.0 : 1.0)
+                                .foregroundColor(Color.theme.blue)
+                                .onTapGesture {
+                                    viewModel.fetchData(queryType: .busesByNumber(search: searchText))
+                                }
+                            
+                            Image(systemName: "xmark.circle.fill")
+                                .padding()
+                                .offset(x: 10)
+                                .opacity(searchText.isEmpty ? 0.0 : 1.0)
+                                .foregroundColor(Color.theme.blue)
+                                .onTapGesture {
+                                    UIApplication.shared.endEditing()
+                                    searchText = ""
+                                    viewModel.showBusList = false
+                                }
+                        }
+                        , alignment: .trailing
+                    )
             }
-            .onEnded { val in
-                prevDragTranslation = .zero
-                isDragging = false
-                if currentHeight > maxHeight {
-                    currentHeight = maxHeight
-                }else if currentHeight < minHeight{
-                    currentHeight = minHeight
+            .padding(.horizontal)
+            .padding(.top, 64)
+            Spacer()
+            
+            Divider()
+                .padding(.vertical)
+            
+            if viewModel.showBusList {
+                if let buses = viewModel.buses as [Bus]? {
+                    List(buses, id: \.gtfsId) { bus in
+                        Button {
+                            viewModel.fetchData(queryType: .routeByBus(search: bus.shortName))
+                            viewModel.showBusList = false
+                            viewModel.showDirectionList = true
+                            viewModel.selectedBus = bus.shortName
+                        } label: {
+                            Text(bus.shortName)
+                        }
+                    }
+                    .listStyle(PlainListStyle())
                 }
             }
+            
+            if viewModel.showDirectionList {
+                if let patternArray = viewModel.patternArray as [Pattern]? {
+                    List(patternArray, id: \.name) { item in
+                        NavigationLink(
+                            item.name,
+                            destination:
+                                MapViewPassenger(
+                                    patternGeometry: item.patternGeometry.points,
+                                    stopsInfo: viewModel.stops,
+                                    selectedBus: viewModel.selectedBus)
+                                .ignoresSafeArea()
+                            
+                        )
+                    }
+                    .listStyle(PlainListStyle())
+                    .padding()
+                }
+            }
+        }
+        .background(Color.theme.background)
+        
     }
 }
-
-struct SearchPassangerView_Previews: PreviewProvider {
+struct SearchPassengerView_Previews: PreviewProvider {
     static var previews: some View {
-        SearchPassengerView(showSearchScreen: .constant(true))
-        // PassengerView()
+        SearchPassengerView(showSearchView: .constant(true))
     }
 }
-
